Solving ODEs with Euler's Method
Overview
This project focuses on solving Ordinary Differential Equations (ODEs) using Euler's Method, a fundamental numerical method in the field of numerical computing and engineering. Aimed at demonstrating the application of Euler's Method to model physical systems, this project provides an in-depth exploration of solving first-order ODEs, critical in various natural sciences and mathematics contexts.

Project Description
In this lab, we've developed a script to solve a specific ODE that models the motion of a mass. The ODE given is y' + y * cos(t) = 0, with an initial value of y(0) = 1.241. Our goal was to find the values of displacement for 0 ≤ t ≤ 6 using different step sizes (h = 0.5, 0.25, 0.1) and to compare these numerical solutions with the exact solution, y = 0.5 * e^(sin(t)) * 2e^(-sin(t)), calculating both absolute and relative errors for h = 0.5.

Features
Implementation of Euler's Method to solve first-order ODEs.
Calculation of numerical solutions for varying step sizes.
Comparison of numerical solutions with the analytical solution.
Error analysis including both absolute and relative errors.
Technologies Used
R Programming Language: Utilized for all computations and plotting the results for visualization.
How to Run
To run this project, ensure you have R installed on your machine. Open the provided .R script in an R environment (e.g., RStudio), and execute the script. The program will output plots for each step size and the error analysis.

Contribution
This project is part of the coursework for CST8233 - Numerical Computing at the School of Advanced Technology. Contributions and improvements are welcome to refine the calculations or enhance the visualization of results.